Expanded left side over input field when cannot type, combined elements:
Instead of four list variables for the different usages, there are
now only two (to have different SetDoneFunc behavior).
Instead of five forms there are now two, instead of 11 text boxes there
are now four, and instead of two input fields there is now one.
It is more consistent now that you must write a displayName to save a
field after editing it in /new.
Can now press esc to go back to the top leaving the fields added list
in /new.